+++ src/shared/layouts/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router';
 
-import { Header } from '../components/ui/Header';
+import { Header } from '../../widgets/Header';
 
 export const Layout = () => {
   return (
