+++ src/widgets/Header.tsx
@@ -3,8 +3,8 @@ import { useMemo } from 'react';
 import '../../../app/global/App.css';
 import { Link } from 'react-router';
 
-import { ROUTES } from '../../../app/routes/routes';
-import logo from '../../../shared/components/logo/Megabox_Logo_Indigo.png';
+import { ROUTES } from '../app/routes/routes';
+import logo from '../shared/components/logo/Megabox_Logo_Indigo.png';
 
 export const Header = () => {
   const today = useMemo(() => {
